add project files
