+++ .github/ISSUE_TEMPLATE.md
@@ -1,0 +1,25 @@
+## Expected Behavior
+
+Describe what you were trying to get done and what you expected to happen.
+
+## Actual Behavior
+
+Tell us what happened, what went wrong,
+
+## Steps to Reproduce the Problem
+
+  1.  
+  2.  
+  3.  
+
+```plaintext
+Paste the command(s) you ran and the output.
+If there was a crash, please include the traceback here.
+```
+
+## Specifications
+
+* Python version:
+* GINO version:
+* gino-quart version:
+* Quart version:
