+++ README.md
@@ -1,0 +1,13 @@
+# gino-quart
+
+An extension for GINO to support Quart server.
+
+**This project "gino-quart" is currently not maintained and needs adoption.** 
+
+Since GINO 1.0, the built-in extensions are now separate projects supported by
+the community. This project is copied here directly from GINO 0.8.x for
+compatibility. Help is needed to:
+
+* Keep this project maintained - follow Quart releases, fix issues, etc.
+* Add more examples and documentation.
+* Answer questions in the community.
