+++ src/gino_quart.py
@@ -1,15 +1,11 @@
 import asyncio
 
-# noinspection PyPackageRequirements
+from gino.api import Gino as _Gino, GinoExecutor as _Executor
+from gino.engine import GinoConnection as _Connection, GinoEngine as _Engine
+from gino.strategies import GinoStrategy
 from quart import Quart, request
-
-# noinspection PyPackageRequirements
 from quart.exceptions import NotFound
 from sqlalchemy.engine.url import URL
-
-from ..api import Gino as _Gino, GinoExecutor as _Executor
-from ..engine import GinoConnection as _Connection, GinoEngine as _Engine
-from ..strategies import GinoStrategy
 
 
 class QuartModelMixin:
